Create routes for schedules and update entity routes to integrate with schedules

+++ src/controllers/entityController.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 import { Group } from "../models/Group.js";
 import { Location } from "../models/Location.js";
+import { Schedule } from "../models/Schedule.js";
 import { Teacher } from "../models/Teacher.js";
 import { Entity, EntityCategory } from "../types.js";
 import { ErrorCode, SuccessCode } from "../utils/constants.js";
@@ -9,6 +10,7 @@ import { ErrorResponse, sendValidResponse } from "../utils/sendResponse.js";
 type CreateEntityBody = {
   category: EntityCategory;
   displayName: string;
+  schedule: string;
 };
 
 type GetEntitiesQuery = {
@@ -18,31 +20,42 @@ type GetEntitiesQuery = {
 type DeleteEntityQuery = {
   id: string;
   category: EntityCategory;
+  schedule: string;
 };
 
 async function createEntity(
   req: Request<{}, { id: string }, CreateEntityBody, {}>,
   res: Response,
 ) {
-  const { displayName, category } = req.body;
+  const { displayName, category, schedule } = req.body;
   let result;
+
+  const findSchedule = await Schedule.findById(schedule);
+  if (!findSchedule) {
+    throw new ErrorResponse(ErrorCode.BAD_REQUEST, "Invalid schedule.");
+  }
+
   try {
     switch (category) {
       case EntityCategory.TEACHER:
         const teacherResult = await Teacher.create({ displayName });
         result = teacherResult;
+        findSchedule.teachers.push(teacherResult._id);
         break;
       case EntityCategory.GROUP:
         const groupResult = await Group.create({ displayName });
         result = groupResult;
+        findSchedule.groups.push(groupResult._id);
         break;
       case EntityCategory.LOCATION:
         const locationResult = await Location.create({ displayName });
         result = locationResult;
+        findSchedule.locations.push(locationResult._id);
         break;
       default:
         throw new ErrorResponse(ErrorCode.BAD_REQUEST, "Invalid category.");
     }
+    await findSchedule.save();
   } catch (error) {
     console.error(error);
     throw new ErrorResponse(ErrorCode.SERVER_ERROR, "Something went wrong.");
@@ -125,18 +138,28 @@ async function deleteEntity(
   req: Request<{}, void, {}, DeleteEntityQuery>,
   res: Response,
 ) {
-  const { id, category } = req.query;
+  const { id, category, schedule } = req.query;
+
+  const findSchedule = await Schedule.findById(schedule);
+  if (!findSchedule) {
+    throw new ErrorResponse(ErrorCode.BAD_REQUEST, "Invalid schedule.");
+  }
 
   try {
     switch (category) {
       case EntityCategory.TEACHER:
         await Teacher.deleteOne({ _id: id });
+        await Schedule.findByIdAndUpdate(schedule, { $pull: { teachers: id } });
         break;
       case EntityCategory.GROUP:
         await Group.deleteOne({ _id: id });
+        await Schedule.findByIdAndUpdate(schedule, { $pull: { groups: id } });
         break;
       case EntityCategory.LOCATION:
         await Location.deleteOne({ _id: id });
+        await Schedule.findByIdAndUpdate(schedule, {
+          $pull: { locations: id },
+        });
         break;
       default:
         throw new ErrorResponse(ErrorCode.BAD_REQUEST, "Invalid category.");
